Refactor constant definition parsing into dedicated defines module
The parseDefine* functions didn't fit well in mounts.ts since they handle constant definitions, not filesystem mounts. Moved all constant parsing and merging logic into a new defines.ts module to better organize the codebase.

Changes:
- Created src/defines.ts with all constant parsing functions
- Extracted common mergeConstants() helper to eliminate duplicate logic between mergeConstantsForThisRun and mergeConstantsForWpConfig
- Removed parseDefineStringArguments, parseDefineBoolArguments, parseDefineNumberArguments from mounts.ts
- Updated imports across the codebase to use the new defines module
- Removed unnecessary `as string` type casts in boot.ts since php.defineConstant() already accepts the correct union type
